Load dark mode stylesheet early on
Previously, if the user switched pages while in dark mode, the page
would briefly flash in regular mode before switching to dark mode. This
happened because the dark mode script (which switches the stylesheet)
ran after the whole document loaded. To fix this, we moved the
initial mode loading to a script in the head of the document.

We also made a slight change to the style of the toggle switch.